async support feature-gated
