+++ mirrord/console/Cargo.toml
@@ -19,10 +19,11 @@ required-features = ["binary"]
 
 [features]
 default = []
-binary = ["dep:tracing", "dep:tracing-subscriber"]
+binary = ["dep:tracing", "dep:tracing-subscriber", "async-logger"]
+async-logger = ["mirrord-intproxy-protocol/codec-async"]
 
 [dependencies]
-mirrord-intproxy = { path = "../intproxy" }
+mirrord-intproxy-protocol = { path = "../intproxy/protocol", features = ["codec"] }
 tokio.workspace = true
 tracing = { workspace = true, optional = true }
 tracing-subscriber = { workspace = true, optional = true }
