+++ mirrord/layer/Cargo.toml
@@ -19,7 +19,6 @@ mirrord-config = { path = "../config"}
 mirrord-protocol = { path = "../protocol"}
 mirrord-layer-macro = { path = "./macro"}
 mirrord-console = { path = "../console" }
-mirrord-intproxy = { path = "../intproxy" }
 mirrord-intproxy-protocol = { path = "../intproxy/protocol" }
 
 ctor = "0.2"
